# openapi.yaml
...
openapi: 3.0.3
info:
  title: LoanFlow API
  description: |
    # LoanFlow - Fintech Loan Management Platform API
    
    This API provides endpoints for managing loans, users, repayments, 
    and system health monitoring.

    ## Foundation Phase Features
    - System health checks
    - API connectivity testing
    - Monitoring and diagnostics

    ## Homepage Phase Features
    - Customer testimonials
    - Frequently Asked Questions
    - Loan products
    - Settings (hero, statistics, features)

    ## About Page Phase Features
    - Get and manage Team Members
    - Get and update About Page settings (Mission, Story, Values)
    
    ## Upcoming Features
    - User authentication
    - Loan application processing
    - Repayment management
    - Admin dashboard
    
  version: 1.2.0
  contact:
    name: LoanFlow Support
    email: support@loanflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development Server
  - url: https://loanflow-backend-fprb.onrender.com
    description: Production Server
...
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        _id:
          type: string
          example: 67a3b4c5d6e7f8g9h0i1j2k9
        name:
          type: string
          example: Adedayo Williams
        role:
          type: string
          example: CEO & Co-Founder
        bio:
          type: string
          example: Adedayo has over 15 years of experience in fintech and banking...
        photoUrl:
          type: string
          nullable: true
        photoPublicId:
          type: string
          nullable: true
        order:
          type: integer
          default: 0
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AboutSettings:
      type: object
      properties:
        _id:
          type: string
        heroHeading:
          type: string
          example: Empowering Financial Freedom
        heroSubheading:
          type: string
          example: We're on a mission to make loans accessible...
        heroImageUrl:
          type: string
          nullable: true
        heroImagePublicId:
          type: string
          nullable: true
        storyTitle:
          type: string
          example: Our Story
        storyContent:
          type: string
          example: LoanFlow was founded in 2020...
        missionStatement:
          type: string
          example: To democratize access to affordable loans...
        values:
          type: array
          items:
            $ref: '#/components/schemas/Value'
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Value:
      type: object
      properties:
        title:
          type: string
          example: Accessibility
        description:
          type: string
          example: Making loans available to everyone...
        icon:
          type: string
          example: UserGroupIcon
paths:
...
  /api/team:
    get:
      tags:
        - Team
      summary: Get Team Members
      description: Retrieves active team members
      operationId: getTeamMembers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of team members to return
        - name: isActive
          in: query
          schema:
            type: boolean
            default: true
          description: Filter by active status
      responses:
        '200':
          description: Team members retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Team members retrieved successfully
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamMember'
                  timestamp:
                    type: string
                    format: date-time
  /api/team/{id}:
    get:
      tags:
        - Team
      summary: Get Single Team Member
      description: Retrieves a specific team member by ID
      operationId: getTeamMemberById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Team member ID
      responses:
        '200':
          description: Team member retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Team member retrieved successfully
                  data:
                    $ref: '#/components/schemas/TeamMember'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Team member not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Team member not found
                  timestamp:
                    type: string
                    format: date-time
  /api/settings/about:
    get:
      tags:
        - Settings
      summary: Get About Settings
      description: Retrieves about page settings
      operationId: getAboutSettings
      responses:
        '200':
          description: About settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: About settings retrieved successfully
                  data:
                    $ref: '#/components/schemas/AboutSettings'
                  timestamp:
                    type: string
                    format: date-time

    put:
      tags:
        - Settings
      summary: Update About Settings
      description: Updates about page settings
      operationId: updateAboutSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                heroHeading:
                  type: string
                heroSubheading:
                  type: string
                heroImageUrl:
                  type: string
                heroImagePublicId:
                  type: string
                storyTitle:
                  type: string
                storyContent:
                  type: string
                missionStatement:
                  type: string
                values:
                  type: array
                  items:
                    $ref: '#/components/schemas/Value'
      responses:
        '200':
          description: About settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: About settings updated successfully
                  data:
                    $ref: '#/components/schemas/AboutSettings'
                  timestamp:
                    type: string
                    format: date-time